(Willian Filipe 31/07/2025) Feat: Iniciada integração com a tomada inteligente tuya smart

+++ Entities/Registro.cs
@@ -7,6 +7,6 @@ namespace HyperEfficient.Entities
         public DateTime DataInicial { get; set; }
         public DateTime? DataFinal { get; set; }
         public int EquipamentoId { get; set; }
-        public decimal? TotalTempo { get; private set; }
+        public double? TotalTempo { get; private set; }
     }
 }
